test(api): split ContentFunctionsTests by subject
The file had become a grab-bag: 1166 lines covering blog, events, resources,
newsletters, me/whoami and subscribers, with its section banners drifted out of
step with the contents after successive splices. Nothing in it tested a
"ContentFunctions" class, because no such class exists.

Split by subject into six files, routed by the factory each test exercises rather
than by line range — the banners could not be trusted. The cross-surface smoke
test that walks the newsletters surface and finishes with a subscribe goes with
newsletters and borrows the subscribers factory.

Pure move: 376 tests before, 376 after. That count is the review check.

Done first because the next commits add mutation handlers whose tests belong in a
file called ContentFunctionsTests.cs — a name this file was occupying without
earning.
